requirements.txt
finnhub-python==2.4.20
Flask==3.0.3
requests==2.31.0
langchain-openai==0.1.8
langchain-core==0.2.7
